% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{cutoff}
\alias{cutoff}
\title{Extract the optimal cutoff.}
\usage{
cutoff(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to the methods.}
}
\description{
Extracts the optimal cutoff from an object.
}
