% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{cutoff.survcut}
\alias{cutoff.survcut}
\title{The optimal cutoff.}
\usage{
\method{cutoff}{survcut}(x, ...)
}
\arguments{
\item{x}{a survcut object.}

\item{...}{extra arguments, currently none.}
}
\value{
a numeric vector with the optima cutoffs.
}
\description{
Extracts the optimal cutoff from a survcut object.
}
