% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{formula.survcut}
\alias{formula.survcut}
\title{Formula of a survcut object.}
\usage{
\method{formula}{survcut}(x, ...)
}
\arguments{
\item{x}{a survcut object.}

\item{...}{extra arguments, currently none.}
}
\value{
a formula or a list of formulas, if more optimal, cutoffs were found.
}
\description{
Extracts the formula of a survcut object, which may be used
survival difference testing.
}
