% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{is_survcut}
\alias{is_survcut}
\title{Checks for the survcut class.}
\usage{
is_survcut(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments, currently none.}
}
\value{
a logical value.
}
\description{
Checks if the object is an instance of the survcut class.
}
