% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isValidName}
\alias{isValidName}
\title{Check if the string needs non-standard name escape.}
\usage{
isValidName(x)
}
\arguments{
\item{x}{a string vector.}
}
\value{
a logical value.
}
\description{
Checks if a string is a syntactically valid name.
}
