% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{model.frame.survcut}
\alias{model.frame.survcut}
\title{Stratified data set.}
\usage{
\method{model.frame}{survcut}(formula, ...)
}
\arguments{
\item{formula}{a survcut object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the stratified variables.
}
\description{
Extracts the data frame with the stratified data set from
a survcut object.
}
