% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{plot.survcut}
\alias{plot.survcut}
\title{Plot a survcut object.}
\usage{
\method{plot}{survcut}(
  x,
  type = c("km", "diagnostic"),
  palette = c("steelblue", "firebrick"),
  ggtheme = survminer::theme_survminer(),
  ...
)
}
\arguments{
\item{x}{a survcut object.}

\item{type}{plot type: 'km' plots a Kaplan-Meier plot, 'diagnostic' generates
optimization plots.}

\item{palette}{a color palette used in the Kaplan-Meier plot.}

\item{...}{extra arguments passed to \code{\link[survminer]{ggsurvplot}}.}
}
\value{
a ggplot or a list of ggplots.
}
\description{
Plots Kaplan-Meier curves for the variable strata or a series
of diagnostic plots visualizing the optimization process.
}
\details{
For Kaplan-Meier plots, a wrapper around
\code{\link[survminer]{ggsurvplot}}. Cutoff and p values are presented in the
plot subtitle, complete observation counts are shown in the plot tag.
If more cutoffs were identified, returns a list with the Kaplan-Meier plots.
}
