% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{print.survcut}
\alias{print.survcut}
\title{Appearance of the survcut object.}
\usage{
\method{print}{survcut}(x, ...)
}
\arguments{
\item{x}{a survcut object.}

\item{...}{extra arguments, currently none.}
}
\value{
none, called for it's side effects.
}
\description{
Prints a survcut object.
}
