% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{summary.survcut}
\alias{summary.survcut}
\title{Testing summary of a survcut object.}
\usage{
\method{summary}{survcut}(object, type = c("cutoff", "all"), ...)
}
\arguments{
\item{object}{a survcut object.}

\item{type}{type of the returned summary: 'all' returns the entire results,
'cutoff' (default) returns the stats for the optimal cutoffs.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the testing stats.
}
\description{
Returns the testing result summary for a survcut object.
}
