% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{surv_diff}
\alias{surv_diff}
\title{Test for survival differences between the strata.}
\usage{
surv_diff(x, ...)
}
\arguments{
\item{x}{a survcut object.}

\item{...}{extra arguments passed to \code{\link[survival]{survdiff}}.}
}
\value{
a survdiff object or a list of survdiff objects.
}
\description{
Tests for differences in survival for the variable stratified by
the optimal cutoffs.
}
\details{
Technically, a wrapper around \code{\link[survival]{survdiff}}.
If more cutoffs were identified, returns a list with the survdiff objects.
}
