% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{surv_fit}
\alias{surv_fit}
\title{Survival curves for a survcut object.}
\usage{
surv_fit(x, ...)
}
\arguments{
\item{x}{a survcut object.}

\item{...}{extra arguments passed to \code{\link[survival]{survfit}}.}
}
\value{
a survfit object or a list of survfit objects.
}
\description{
Calculates survival curves for the variable stratified by
the optimal cutoffs.
}
\details{
Technically, a wrapper around \code{\link[survival]{survfit}}.
If more cutoffs were identified, returns a list with the survdfit objects.
}
