% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{survcut}
\alias{survcut}
\title{Generate a survcut object.}
\usage{
survcut(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments, currently none specified.}
}
\description{
Creates a survcut object on a top of a named list with the
'cutoff, 'cutoff_stats', 'data' and 'test' components.
}
