% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{survfit.default}
\alias{survfit.default}
\title{Create survival curves.}
\usage{
survfit.default(formula, ...)
}
\arguments{
\item{formula}{either a formula or a previously fitted model.}

\item{...}{extra arguments passed to the methods.}
}
\value{
An object of class survfit containing one or more survival curves.
}
\description{
A default method, creates survival curves from a formula or
an object with the survival modeling results.
}
