% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_cutoff}
\alias{test_cutoff}
\title{Test survival differences at the variable cutoff.}
\usage{
test_cutoff(data, time, event, variable, cutoff, ...)
}
\arguments{
\item{time}{name of the survival time variables, needs to be numeric.}

\item{event}{name of the binary event variable: 1 is event, 0 no event.}

\item{variable}{name of the variable to be cut, needs to be numeric.}

\item{cutoff}{the cutoff value.}

\item{...}{additional arguments passed to \code{\link[survival]{survdiff}}.}
}
\value{
a data frame with the cutoff value, chi-squared test statistic,
degrees of freedom, p value and n numbers in the low and high strata.
}
\description{
Tests for the differences in survival for observations
stratified by the given variable cutoff. The testing ia accomplished by
the \code{\link[survival]{survdiff}} function.
}
